Refactor multipart
